fix(Modal): 모달의 디렉토리 위치 변경 및 기능 축소 기본 모달은 헤더만 필수로 보여주고 나머지는 children 으로 구현하도록 기능 축소
