+++ components/src/lib/Modal/Modal/Modal.stories.tsx
@@ -1,56 +1,44 @@
 import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
-import Modal from "../Modal/common/Modal";
+import Modal from "./Modal";
 
 const meta = {
-  title: "ModalExample",
+  title: "Modal/Default",
   component: Modal,
   parameters: {
     docs: {
       description: {
-        component: "모달 컴포넌트",
+        component: "기본 모달 컴포넌트",
       },
     },
   },
   argTypes: {
+    modalTitle: {
+      description: "모달 헤더의 텍스트",
+    },
     position: {
       description: "modal의 위치",
     },
     modalSize: {
       description: "modal의 크기",
     },
-    closeButton: {
-      description: "modal을 닫는 closeButton의 형태",
-    },
     closeModal: {
       description: "modal을 닫을 떄의 action",
     },
-    title: {
-      description: "modal의 제목",
-    },
     children: {
       description: "modal의 내용",
-    },
-    buttonText: {
-      description: "(optional) button의 text",
-    },
-    buttonClick: {
-      description: "(optional) 버튼을 클릭했을 때 action",
     },
   },
 
   decorators: [
     (Story, context) => {
       const [openModal, setOpenModal] = React.useState(false);
-      const clickModalButton = () => alert("버튼 클릭 성공");
       const closeModal = () => setOpenModal(false);
 
       return (
         <div>
           <button onClick={() => setOpenModal(true)}>모달 열기</button>
-          {openModal && (
-            <Story args={{ ...context.args, closeModal, buttonClick: clickModalButton }} />
-          )}
+          {openModal && <Story args={{ ...context.args, closeModal }} />}
         </div>
       );
     },
@@ -65,11 +53,8 @@ export const CenterImgModal: Story = {
   args: {
     position: "center",
     children: "칠드런",
-    closeButton: "img",
     closeModal: () => {},
-    title: "제목임",
-    buttonText: "버튼임",
-    buttonClick: () => {},
+    modalTitle: "기본 모달이에용",
   },
 };
 
@@ -78,10 +63,7 @@ export const BottomTextModal: Story = {
     position: "bottom",
     children: "칠드런",
     modalSize: "m",
-    closeButton: "text",
     closeModal: () => {},
-    title: "제목임",
-    buttonText: "버튼임",
-    buttonClick: () => {},
+    modalTitle: "제목임",
   },
 };
