--- components/src/lib/Modal/common/Modal.css
@@ -1,123 +1,0 @@
-.darr-modal {
-  position: fixed;
-  width: 100vw;
-  height: 100vh;
-  top: 0;
-  left: 0;
-  color: black;
-}
-
-.darr-modal .modal-backdrop {
-  background-color: rgba(0, 0, 0, 0.35);
-  width: 100vw;
-  height: 100vh;
-  z-index: 1;
-}
-
-.darr-modal .modal-inner {
-  position: absolute;
-  background-color: white;
-  height: fit-content;
-  z-index: 2;
-  border-radius: 8px;
-  padding: 24px 32px;
-  box-sizing: border-box;
-  max-height: 60%;
-  overflow: auto;
-}
-
-.darr-modal .modal-inner.center {
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  min-height: 216px;
-}
-
-.darr-modal .size-s {
-  width: 40vw;
-}
-
-.darr-modal .size-m {
-  width: 60vw;
-}
-
-.darr-modal .size-l {
-  width: 80vw;
-}
-
-.darr-modal .modal-inner.bottom {
-  width: 100vw;
-  min-height: 206px;
-  bottom: 0;
-}
-
-.darr-modal .modal-inner .modal-top {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 24px;
-}
-.darr-modal .modal-inner .modal-top .modal-title {
-  font-size: 18px;
-  font-weight: 700;
-  line-height: 26px;
-}
-
-.darr-modal .modal-inner .modal-top .modal-close-btn {
-  background-color: transparent;
-  padding: 0;
-  border: none;
-  outline: none;
-  cursor: pointer;
-}
-.darr-modal .modal-top .modal-close-btn .modal-close-img {
-  width: 14px;
-  height: 14px;
-}
-
-.darr-modal .modal-inner .modal-content {
-  width: 100%;
-  height: 100%;
-  min-height: 50px;
-  margin-bottom: 16px;
-}
-
-.darr-modal .modal-inner .modal-bottom {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-}
-
-.darr-modal .modal-bottom .modal-confirm-btn {
-  background-color: rgba(51, 51, 51, 1);
-  color: white;
-  padding: 10px 0;
-  font-size: 15px;
-  font-weight: 700;
-  box-sizing: border-box;
-  border: none;
-  border-radius: 8px;
-  outline: none;
-  cursor: pointer;
-}
-
-.darr-modal .modal-bottom .modal-confirm-btn:hover {
-  background-color: black;
-}
-
-.darr-modal .modal-bottom .modal-close-text-btn {
-  background-color: transparent;
-  color: rgba(139, 149, 161, 1);
-  padding: 10px 0;
-  font-size: 15px;
-  font-weight: 700;
-  box-sizing: border-box;
-  border: none;
-  outline: none;
-  cursor: pointer;
-}
-
-.darr-modal .modal-bottom .modal-close-text-btn:hover {
-  color: black;
-  border: none;
-  outline: none;
-}
